방장 입금계좌 설정 로직, 신고하기 기능 추가

+++ src/main/java/com/example/spendolive/report/domain/ReportVO.java
@@ -8,7 +8,7 @@ public class ReportVO {
     private Long report_id;             // NUMBER -> Long
     private String reporter_id;         // VARCHAR2(20)
     private String reported_member_id;  // VARCHAR2(20)
-    private Long room_id;               // NUMBER -> Long (Null 허용이므로 래퍼 클래스 사용)
+    private int room_id;               // NUMBER -> Long (Null 허용이므로 래퍼 클래스 사용)
     private String report_reason;       // VARCHAR2(500)
     private String report_status;       // VARCHAR2(20) DEFAULT 'WAIT'
     private String admin_comment;       // VARCHAR2(1000)
